refactor table reading tests (refs #110)
