+++ validator-core/src/test/java/fr/ign/validator/repository/xml/XmlDocumentModelRepositoryTest.java
@@ -2,43 +2,44 @@ package fr.ign.validator.repository.xml;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Collections;
 import java.util.List;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import fr.ign.validator.tools.ResourceHelper;
 import fr.ign.validator.model.DocumentModel;
-
+import fr.ign.validator.tools.ResourceHelper;
 
 public class XmlDocumentModelRepositoryTest {
 
-	private File configDir ;
-		
+	private File configDir;
+
 	@Before
 	public void setUp() {
-		configDir = ResourceHelper.getResourceFile(getClass(),"/config");
+		configDir = ResourceHelper.getResourceFile(getClass(), "/config");
 	}
 
 	@Test
-	public void testListAll(){
+	public void testListAll() {
 		XmlDocumentModelRepository repository = new XmlDocumentModelRepository(configDir);
 		List<String> documentModelNames = repository.listAll();
-		Assert.assertEquals(1,documentModelNames.size());
+		Assert.assertEquals(3, documentModelNames.size());
+		Collections.sort(documentModelNames);
 		Assert.assertEquals("cnig_PLU_2014", documentModelNames.get(0));
 	}
-	
+
 	@Test
-	public void testFindOneByName() throws IOException{
+	public void testFindOneByName() throws IOException {
 		XmlDocumentModelRepository repository = new XmlDocumentModelRepository(configDir);
 		DocumentModel documentModel = repository.findOneByName("cnig_PLU_2014");
 		Assert.assertNotNull(documentModel);
 		Assert.assertEquals("cnig_PLU_2014", documentModel.getName());
 	}
-	
+
 	@Test
-	public void testFindOneByNameNotFound(){
+	public void testFindOneByNameNotFound() {
 		XmlDocumentModelRepository repository = new XmlDocumentModelRepository(configDir);
 		try {
 			Assert.assertNull(repository.findOneByName("not_found"));
