+++ validator-core/src/test/java/fr/ign/validator/tools/FileConverterTest.java
@@ -2,6 +2,7 @@ package fr.ign.validator.tools;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 
 import org.apache.commons.io.FileUtils;
@@ -12,7 +13,6 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
-import fr.ign.validator.tools.ResourceHelper;
 import fr.ign.validator.tools.ogr.OgrVersion;
 
 /**
@@ -38,21 +38,109 @@ public class FileConverterTest {
 	}
 
 	@Test
-	public void testConvertToCSV() throws IOException {
-		File source = ResourceHelper.getResourceFile(getClass(),"/data/ZONE_URBA_41003.TAB");
-		File target = folder.newFile("ZONE_URBA.csv");
-		try {
-			fileConverter.convertToCSV(source, target);
-			Assert.assertTrue(target.exists());
-			List<String> lines = FileUtils.readLines(target);
-			Assert.assertEquals(37, lines.size());
-			Assert.assertEquals("WKT,LIBELLE,LIBELONG,TYPEZONE,DESTDOMI,NOMFIC,URLFIC,INSEE,DATAPPRO,DATVALID",
-					lines.get(0));
-		} catch (Exception e) {
-			e.printStackTrace();
-			Assert.fail();
-		}
+	public void testConvertShpLatin1ToCSV() throws IOException {
+		File source = ResourceHelper.getResourceFile(getClass(),"/data/shp_latin1/PRESCRIPTION_PCT.dbf");
+		File target = folder.newFile("PRESCRIPTION_PCT.csv");
+		fileConverter.convertToCSV(source, target);
+		Assert.assertTrue(target.exists());
+		List<String> lines = FileUtils.readLines(target,StandardCharsets.ISO_8859_1);
+		Assert.assertEquals(60, lines.size());
+		Assert.assertEquals(
+			"WKT,LIBELLE,TXT,TYPEPSC,NOMFIC,URLFIC,INSEE,DATAPPRO,DATVALID",
+			lines.get(0)
+		);
+		Assert.assertEquals(
+			"\"POINT (557311.155866353 6742836.60415676)\",Bâtiment agricole,,16,,,41269,2013/09/26,",
+			lines.get(1)
+		);
 	}
+	
+	@Test
+	public void testConvertShpUtf8ToCSV() throws IOException {
+		File source = ResourceHelper.getResourceFile(getClass(),"/data/shp_utf8/PRESCRIPTION_PCT.dbf");
+		File target = folder.newFile("PRESCRIPTION_PCT.csv");
+		fileConverter.convertToCSV(source, target);
+		Assert.assertTrue(target.exists());
+		List<String> lines = FileUtils.readLines(target,StandardCharsets.UTF_8);
+		Assert.assertEquals(60, lines.size());
+		Assert.assertEquals(
+			"WKT,LIBELLE,TXT,TYPEPSC,NOMFIC,URLFIC,INSEE,DATAPPRO,DATVALID",
+			lines.get(0)
+		);
+		Assert.assertEquals(
+			"\"POINT (557311.155866353 6742836.60415676)\",Bâtiment agricole,,16,,,41269,2013/09/26,",
+			lines.get(1)
+		);
+	}
+
+	@Test
+	public void testConvertTabLatin1ToCSV() throws IOException {
+		File source = ResourceHelper.getResourceFile(getClass(),"/data/tab_latin1/PRESCRIPTION_PCT.TAB");
+		File target = folder.newFile("PRESCRIPTION_PCT.csv");
+		fileConverter.convertToCSV(source, target);
+		Assert.assertTrue(target.exists());
+		List<String> lines = FileUtils.readLines(target,StandardCharsets.ISO_8859_1);
+		Assert.assertEquals(60, lines.size());
+		Assert.assertEquals(
+			"WKT,LIBELLE,TXT,TYPEPSC,NOMFIC,URLFIC,INSEE,DATAPPRO,DATVALID",
+			lines.get(0)
+		);
+		Assert.assertEquals(
+			"\"POINT (557311.156 6742836.604)\",Bâtiment agricole,,16,,,41269,2013/09/26,",
+			lines.get(1)
+		);
+	}
+	
+	@Test
+	public void testConvertTabUtf8ToCSV() throws IOException {
+		File source = ResourceHelper.getResourceFile(getClass(),"/data/tab_utf8/PRESCRIPTION_PCT.tab");
+		File target = folder.newFile("PRESCRIPTION_PCT.csv");
+		fileConverter.convertToCSV(source, target);
+		Assert.assertTrue(target.exists());
+		List<String> lines = FileUtils.readLines(target,StandardCharsets.UTF_8);
+		Assert.assertEquals(60, lines.size());
+		Assert.assertEquals(
+			"WKT,LIBELLE,TXT,TYPEPSC,NOMFIC,URLFIC,INSEE,DATAPPRO,DATVALID",
+			lines.get(0)
+		);
+		Assert.assertEquals(
+				"\"POINT (557311.17 6742836.6)\",Bâtiment agricole,,16,,,41269,2013/09/26,",
+			lines.get(1)
+		);
+	}
+	
+	@Test
+	public void testConvertGmlUtf8ToCSV() throws IOException {
+		File source = ResourceHelper.getResourceFile(getClass(),"/gml/PRESCRIPTION_LIN.gml");
+		File target = folder.newFile("PRESCRIPTION_LIN.csv");
+		fileConverter.convertToCSV(source, target);
+		Assert.assertTrue(target.exists());
+		List<String> lines = FileUtils.readLines(target,StandardCharsets.UTF_8);
+		Assert.assertEquals(92, lines.size());
+		Assert.assertEquals(
+			"WKT,gml_id,LIBELLE,TYPEPSC,TYPEPSC2,INSEE,DATAPPRO",
+			lines.get(0)
+		);
+		Assert.assertEquals(
+			"\"MULTILINESTRING ((226546.544185517 6755336.26169925,226707.572437799 6755556.91246038,226751.823845957 6755621.52764606,226758.594781693 6755631.41445423,226780.994538211 6755664.12220385,226790.366741745 6755677.80733694))\",PRESCRIPTION_LIN.2666,Marge de recul,11,11001,56118,20140123",
+			lines.get(1)
+		);
+	}
+	
+	@Test
+	public void testConvertGmlInvalid() throws IOException {
+		File source = ResourceHelper.getResourceFile(getClass(),"/gml/INVALID.gml");
+		File target = folder.newFile("INVALID.csv");
+		fileConverter.convertToCSV(source, target);
+		Assert.assertTrue(target.exists());
+		List<String> lines = FileUtils.readLines(target,StandardCharsets.UTF_8);
+		Assert.assertEquals(1, lines.size());
+		Assert.assertEquals(
+			"header1,header2,header3",
+			lines.get(0)
+		);
+	}
+	
 
 	/**
 	 * Regress test about GDAL coordinate precision while converting to CSV
@@ -62,18 +150,18 @@ public class FileConverterTest {
 	 * @throws IOException
 	 */
 	@Test
-	public void testConvertToCSVPoint() throws IOException {
+	public void testCoordinatePrecision() throws IOException {
 		File source = ResourceHelper.getResourceFile(getClass(),"/data/POINT.shp");
 		File target = folder.newFile("POINT.csv");
 		
 		String expectedTargetName = getExpectedPointName();
 		Assume.assumeNotNull(expectedTargetName);
 		File expectedTarget = ResourceHelper.getResourceFile(getClass(),expectedTargetName);
-		List<String> expectedLines = FileUtils.readLines(expectedTarget);
+		List<String> expectedLines = FileUtils.readLines(expectedTarget,StandardCharsets.UTF_8);
 		try {
 			fileConverter.convertToCSV(source, target);
 			Assert.assertTrue(target.exists());
-			List<String> lines = FileUtils.readLines(target);
+			List<String> lines = FileUtils.readLines(target,StandardCharsets.UTF_8);
 			Assert.assertEquals(11, lines.size());
 			for ( int i = 0; i < lines.size(); i++ ){
 				String expected = expectedLines.get(i);
@@ -117,7 +205,7 @@ public class FileConverterTest {
 		try {
 			fileConverter.convertToCSV(source, target);
 			Assert.assertTrue(target.exists());
-			List<String> lines = FileUtils.readLines(target);
+			List<String> lines = FileUtils.readLines(target,StandardCharsets.UTF_8);
 			Assert.assertEquals(2, lines.size());
 			Assert.assertEquals("URLFIC,INSEE", lines.get(0));
 			Assert.assertEquals(
@@ -146,7 +234,7 @@ public class FileConverterTest {
 			/* dbf -> csv */
 			fileConverter.convertToCSV(targetDbf, targetCsv);
 			Assert.assertTrue(targetCsv.exists());
-			List<String> lines = FileUtils.readLines(targetCsv);
+			List<String> lines = FileUtils.readLines(targetCsv,StandardCharsets.UTF_8);
 			Assert.assertEquals(2, lines.size());
 			Assert.assertEquals("URLFIC,INSEE", lines.get(0));
 			Assert.assertEquals(
