+++ validator-core/src/test/java/fr/ign/validator/tools/FileUtilsTest.java
@@ -13,7 +13,7 @@ public class FileUtilsTest {
 
 	@Test
 	public void testListFilesAndDirs(){
-		File directory = ResourceHelper.getResourceFile(getClass(),"/geofla") ;
+		File directory = ResourceHelper.getResourceFile(getClass(),"/config/geofla") ;
 		String[] extensions = {"xml"};
 		Collection<File> files = FileUtils.listFilesAndDirs(directory,extensions);
 		assertEquals(4,files.size());
