+++ validator-core/src/test/java/fr/ign/validator/validation/attribute/FilenameExistsValidatorTest.java
@@ -25,7 +25,7 @@ public class FilenameExistsValidatorTest {
 		validator = new FilenameExistsValidator();
 		
 		context = new Context();
-		File currentDirectory = ResourceHelper.getResourceFile(getClass(),"/geofla") ;
+		File currentDirectory = ResourceHelper.getResourceFile(getClass(),"/config/geofla") ;
 		context.setCurrentDirectory(currentDirectory);
 		
 		report = new InMemoryReportBuilder() ;
