+++ validator-core/src/test/resources/config/geofla/types/COMMUNE.xml
