+++ validator-core/src/test/resources/config/geofla/types/LIMITE_COMMUNE.xml
