+++ validator-core/src/test/resources/config/sample-document/files.xml
