+++ validator-core/src/test/resources/config/sample-document/types/COMMUNE.xml
