+++ validator-core/src/test/resources/config/sample-document/types/SIMPLE.xml
