+++ validator-core/src/test/resources/data/tab_latin1/ZONE_URBA_41003.ID
