+++ validator-core/src/test/resources/gml/INVALID.gml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<features>
+<!-- invalid GML to test ogr2ogr behavior !-->
+</features>
+
