--- validator-core/src/test/resources/js/test-strlen-a.js
@@ -1,1 +1,0 @@
-a.length ;
